+++ rails/app/helpers/workspaces_helper.rb
@@ -8,6 +8,11 @@ module WorkspacesHelper
   def update_workspace_from_hash(id, hash)
     workspace = Kaiju::Workspace.by_id(id)
     workspace.name = hash['name'] unless hash['name'].nil?
+    unless hash['component'].nil?
+      update_component(
+        workspace, Kaiju::Component.by_id(workspace.component.value), hash['component']
+      )
+    end
     workspace
   end