+++ changelog.md
@@ -1,6 +1,10 @@
-## January 9th, 2018
+## January 11th, 2018
+### Added
+- Ability to duplicate workspaces from the action bar
+
 ### Fixed
 - The component drop down search is no longer case sensitive
+- Workspace count on project cards
 
 ## January 5th, 2018
 ### Added